allocator: Move custom HeapReport into servo-allocator
The profiler crate shouldn't need to depend on tikv-jemalloc crates.
Instead it makes more sense that the allocator crate provides the
relevant reports to the profiler crate.
