+++ components/profile/Cargo.toml
@@ -31,6 +31,3 @@ regex = { workspace = true }
 
 [target.'cfg(not(target_os = "windows"))'.dependencies]
 libc = { workspace = true }
-
-[target.'cfg(not(any(target_os = "windows", target_env = "ohos")))'.dependencies]
-tikv-jemalloc-sys = { workspace = true }
